Rename format.py to format_data.py because format is an inbuilt fucntion
